ble middleware refactored in order to remove dependency torwards schi from gap/gatt layers. All dependecies with schi were moved to transport layer
