+++ hal_st/middlewares/ble_middleware/GapPeripheralSt.cpp
@@ -11,8 +11,8 @@ namespace
 
 namespace hal
 {
-    GapPeripheralSt::GapPeripheralSt(hal::HciEventSource& hciEventSource, BleBondStorage bleBondStorage, const Configuration& configuration)
-        : GapSt(hciEventSource, bleBondStorage, configuration)
+    GapPeripheralSt::GapPeripheralSt(hal::HciEventSource& hciEventSource, services::BondStorageSynchronizer& bondStorageSynchronizer, const Configuration& configuration)
+        : GapSt(hciEventSource, bondStorageSynchronizer, configuration)
     {
         Initialize(configuration.gapService);
     }
