+++ hal_st/middlewares/ble_middleware/SystemTransportLayer.hpp
@@ -5,7 +5,9 @@
 #include "hal_st/middlewares/ble_middleware/HciEventObserver.hpp"
 #include "infra/util/Function.hpp"
 #include "infra/util/InterfaceConnector.hpp"
+#include "infra/util/ProxyCreator.hpp"
 #include "services/ble/BondBlobPersistence.hpp"
+#include "services/ble/BondStorageSynchronizer.hpp"
 
 namespace hal
 {
@@ -14,6 +16,18 @@ namespace hal
         , public HciEventSource
     {
     public:
+        enum class RfWakeupClock : uint8_t
+        {
+            highSpeedExternal,
+            lowSpeedExternal,
+        };
+
+        struct Configuration
+        {
+            uint16_t maxAttMtuSize;
+            RfWakeupClock rfWakeupClock;
+        };
+
         struct Version
         {
             uint8_t firmwareMajor;
@@ -27,7 +41,9 @@ namespace hal
         };
 
     public:
-        SystemTransportLayer(services::ConfigurationStoreAccess<infra::ByteRange> flashStorage, const infra::Function<void(uint32_t*)>& protocolStackInitialized);
+        using BondStorageSynchronizerCreator = infra::CreatorBase<services::BondStorageSynchronizer, void()>;
+
+        SystemTransportLayer(services::ConfigurationStoreAccess<infra::ByteRange> flashStorage, BondStorageSynchronizerCreator& bondStorageSynchronizerCreator, Configuration configuration, const infra::Function<void(services::BondStorageSynchronizer&)>& onInitialized);
 
         Version GetVersion() const;
 
@@ -53,7 +69,9 @@ namespace hal
 
     private:
         services::BondBlobPersistence bondBlobPersistence;
-        infra::Function<void(uint32_t*)> protocolStackInitialized;
+        infra::DelayedProxyCreator<services::BondStorageSynchronizer, void()> bondStorageSynchronizerCreator;
+        Configuration configuration;
+        infra::AutoResetFunction<void(services::BondStorageSynchronizer&)> onInitialized;
     };
 }
